// back-end/types/ProductOrder.ts
import { BasicCustomer, Customer } from "./Customer";
import { BasicProduct, Product } from "./Product";

export interface BasicOrder {
  product: BasicProduct;
  customer: BasicCustomer;
  productQuantity: number;
}
export interface Order extends BasicOrder {
  orderId: number;
}
export interface OrderWithDetails extends Order {
  product: Product;
  customer: Customer;
}